ci: add per-PR preview deployments to gh-pages
Three workflows render each PR to https://<owner>.github.io/<repo>/pr-preview/pr-N/
so reviewers can see the built UI without checking out the branch. Build runs on
pull_request (fork-safe, read-only token) and uploads an artifact; deploy runs on
workflow_run behind a pr-preview environment that requires maintainer approval,
keeping fork PRs from gaining write access. Dev-fixtures are baked into the build
so the preview renders against the same demo dataset as npm run dev:fixtures.
PWA is disabled in preview builds (VITE_PREVIEW=1) to prevent the production
service worker from intercepting preview navigations.
