+++ .github/workflows/preview-cleanup.yml
@@ -1,0 +1,29 @@
+name: PR Preview Cleanup
+on:
+  pull_request:
+    types: [closed]
+concurrency:
+  group: gh-pages-deploy
+  cancel-in-progress: false
+permissions:
+  contents: write
+  pull-requests: write
+jobs:
+  cleanup:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+        with:
+          ref: gh-pages
+      - name: Remove preview directory
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+          rm -rf "pr-preview/pr-${{ github.event.pull_request.number }}"
+          git add -A
+          git diff --cached --quiet || git commit -m "preview: cleanup PR #${{ github.event.pull_request.number }}"
+          git push
+      - uses: marocchino/sticky-pull-request-comment@v2
+        with:
+          header: pr-preview
+          message: "Preview for this PR has been removed."
