+++ .github/workflows/preview-deploy.yml
@@ -1,0 +1,52 @@
+name: PR Preview Deploy
+on:
+  workflow_run:
+    workflows: ["PR Preview Build"]
+    types: [completed]
+concurrency:
+  group: gh-pages-deploy
+  cancel-in-progress: false
+permissions:
+  contents: write
+  pull-requests: write
+  actions: read
+jobs:
+  deploy:
+    if: github.event.workflow_run.conclusion == 'success'
+    runs-on: ubuntu-latest
+    environment:
+      name: pr-preview
+      url: ${{ steps.url.outputs.url }}
+    steps:
+      - uses: actions/download-artifact@v4
+        with:
+          name: pr-preview-build
+          path: build
+          run-id: ${{ github.event.workflow_run.id }}
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+      - id: pr
+        run: echo "number=$(cat build/.pr-number)" >> "$GITHUB_OUTPUT"
+      - run: rm build/.pr-number
+      - uses: actions/checkout@v6
+        with:
+          ref: gh-pages
+          path: gh-pages
+      - name: Remove stale assets from prior pushes to this PR
+        run: rm -rf "gh-pages/pr-preview/pr-${{ steps.pr.outputs.number }}"
+      - uses: peaceiris/actions-gh-pages@v4
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: ./build
+          destination_dir: pr-preview/pr-${{ steps.pr.outputs.number }}
+          keep_files: true
+          commit_message: "preview: deploy PR #${{ steps.pr.outputs.number }}"
+      - id: url
+        run: echo "url=https://${{ github.repository_owner }}.github.io/${{ github.event.repository.name }}/pr-preview/pr-${{ steps.pr.outputs.number }}/" >> "$GITHUB_OUTPUT"
+      - uses: marocchino/sticky-pull-request-comment@v2
+        with:
+          number: ${{ steps.pr.outputs.number }}
+          header: pr-preview
+          message: |
+            **Preview deployed:** ${{ steps.url.outputs.url }}
+
+            Updated for commit ${{ github.event.workflow_run.head_sha }}.
